api/requirements/prod.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
daphne>=4.0.0,<5.0.0
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-elasticsearch-dsl>=8.0.0,<9.0.0
django-environ==0.12.0
django-phonenumber-field==8.1.0
django-redis>=5.2.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
elasticsearch>=8.0.0,<9.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
phonenumbers==9.0.5
pillow==11.2.1
prompt_toolkit==3.0.51
pytest==8.2.2
pytest-asyncio==0.23.6
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
gunicorn>=20.1.0,<21.0.0
mysqlclient==2.2.4
